vehicle_school/requirements.txt
numpy
tqdm
scikit-learn
scipy>=1.7.3
jax==0.2.27
dm-haiku==0.0.5
optax==0.1.0